Fix MSVC ambiguity when logging the size_t budget constants
The budget diagnostics added upstream stream size_t values into VDebug, which
declares overloads for int32_t, uint32_t, long, unsigned long and double - the
int64_t/uint64_t ones are commented out. On Linux and macOS size_t is
unsigned long, so an exact overload exists and the code builds. On Win64 size_t
is unsigned long long, so there is no exact match and MSVC reports

  error C2593: 'operator <<' is ambiguous

at all three call sites (kMaxLayerNodes once, kMaxShapeContentBudget twice).

Cast to unsigned long at the call sites; the values are small (100000 and
200000) so no truncation is possible on any supported platform. Behaviour is
unchanged: the demo animation still renders RGB(88,214,112) with content
spanning the full banner.

+++ src/lottie/lottieitem.cpp
@@ -106,7 +106,10 @@ static renderer::Layer *createLayerItem(model::Layer *layerData,
         return nullptr;
     }
     if (nodeBudget == 0) {
-        vWarning << "Max render node budget (" << kMaxLayerNodes << ") exceeded";
+        // Cast: VDebug has no size_t overload, and size_t is unsigned long long
+        // on Win64, which makes the call ambiguous under MSVC.
+        vWarning << "Max render node budget ("
+                 << static_cast<unsigned long>(kMaxLayerNodes) << ") exceeded";
         return nullptr;
     }
     --nodeBudget;
@@ -1074,7 +1077,8 @@ void renderer::Group::addChildren(model::Group *data, VArenaAlloc *allocator,
         size_t cost = contentItemCost(*it);
         if (cost > contentBudget) {
             vWarning << "Max shape content budget ("
-                     << kMaxShapeContentBudget << ") exceeded, dropping content item";
+                     << static_cast<unsigned long>(kMaxShapeContentBudget)
+                     << ") exceeded, dropping content item";
             continue;
         }
         contentBudget -= cost;
@@ -1604,7 +1608,8 @@ renderer::Repeater::Repeater(model::Repeater *data, VArenaAlloc *allocator,
 
     for (int i = 0; i < maxCopies; i++) {
         if (contentBudget == 0) {
-            vWarning << "Max shape content budget (" << kMaxShapeContentBudget
+            vWarning << "Max shape content budget ("
+                     << static_cast<unsigned long>(kMaxShapeContentBudget)
                      << ") exceeded, clamping repeater copies to " << i;
             break;
         }
